Extract Migration tests for byron in core-integration
