+++ lib/core-integration/cardano-wallet-core-integration.cabal
@@ -73,6 +73,7 @@ library
       Test.Integration.Scenario.API.Byron.HWWallets
       Test.Integration.Scenario.API.Byron.Addresses
       Test.Integration.Scenario.API.Byron.Transactions
+      Test.Integration.Scenario.API.Byron.Migrations
       Test.Integration.Scenario.API.Shelley.Addresses
       Test.Integration.Scenario.API.Shelley.HWWallets
       Test.Integration.Scenario.API.Shelley.Network
