+++ lib/core-integration/src/Test/Integration/Scenario/API/Shelley/Wallets.hs
@@ -35,7 +35,6 @@ import Cardano.Wallet.Api.Types
     , ApiTransaction
     , ApiUtxoStatistics
     , ApiWallet
-    , ApiWalletMigrationInfo (..)
     , DecodeAddress
     , EncodeAddress (..)
     , WalletStyle (..)
@@ -58,8 +57,6 @@ import Data.Generics.Internal.VL.Lens
     ( view, (^.) )
 import Data.List.NonEmpty
     ( NonEmpty ((:|)) )
-import Data.Maybe
-    ( mapMaybe )
 import Data.Proxy
     ( Proxy (..) )
 import Data.Quantity
@@ -70,11 +67,12 @@ import Data.Text.Class
     ( toText )
 import Data.Word
     ( Word64 )
+import Numeric.Natural
+    ( Natural )
 import Test.Hspec
     ( SpecWith
     , describe
     , it
-    , pendingWith
     , shouldBe
     , shouldNotBe
     , shouldSatisfy
@@ -84,7 +82,6 @@ import Test.Integration.Framework.DSL
     , Headers (..)
     , Payload (..)
     , emptyByronWalletWith
-    , emptyIcarusWallet
     , emptyRandomWallet
     , emptyWallet
     , emptyWalletWith
@@ -95,28 +92,20 @@ import Test.Integration.Framework.DSL
     , expectListSize
     , expectResponseCode
     , expectWalletUTxO
-    , fixtureIcarusWallet
-    , fixturePassphrase
-    , fixtureRandomWallet
     , fixtureWallet
     , getFromResponse
-    , icarusAddresses
     , json
     , listAddresses
     , notDelegating
-    , randomAddresses
     , request
     , selectCoins
     , unsafeRequest
     , verify
     , walletId
-    , (.>)
     , (</>)
     )
 import Test.Integration.Framework.TestData
     ( arabicWalletName
-    , errMsg400ParseError
-    , errMsg403NothingToMigrate
     , errMsg403WrongPass
     , errMsg404NoWallet
     , errMsg406
@@ -138,9 +127,7 @@ import Test.Integration.Framework.TestData
     )
 
 import qualified Cardano.Wallet.Api.Link as Link
-import qualified Cardano.Wallet.Api.Types as ApiTypes
 import qualified Data.List.NonEmpty as NE
-import qualified Data.Map.Strict as Map
 import qualified Data.Text as T
 import qualified Network.HTTP.Types.Status as HTTP
 
@@ -1177,47 +1164,6 @@ spec = do
         expectResponseCode @IO HTTP.status404 ru
         expectErrorMessage (errMsg404NoWallet wid) ru
 
-    describe "BYRON_MIGRATE_05 - I could migrate to any valid address" $ do
-        forM_ [ ("Byron", emptyRandomWallet)
-              , ("Icarus", emptyIcarusWallet)
-              ] $ \(walType, destWallet) -> do
-
-            it ("From wallet type: " ++ walType) $ \ctx -> do
-                --shelley address
-                wShelley <- emptyWallet ctx
-                addrs <- listAddresses @n ctx wShelley
-                let addrShelley = (addrs !! 1) ^. #id
-                --icarus address
-                addrIcarus <- encodeAddress @n . head . icarusAddresses @n
-                    . entropyToMnemonic @15 <$> genEntropy
-                --byron address
-                addrByron <- encodeAddress @n . head . randomAddresses @n
-                    . entropyToMnemonic @12 <$> genEntropy
-
-                sWallet <- destWallet ctx
-                r <- request @[ApiTransaction n] ctx
-                    (Link.migrateWallet @'Byron sWallet)
-                    Default
-                    (Json [json|
-                        { passphrase: #{fixturePassphrase}
-                        , addresses: [#{addrShelley}, #{addrIcarus}, #{addrByron}]
-                        }|])
-                verify r
-                    [ expectResponseCode @IO HTTP.status403
-                    , expectErrorMessage
-                        (errMsg403NothingToMigrate (sWallet ^. walletId))
-                    ]
-
-    it "BYRON_MIGRATE_07 - invalid payload, parser error" $ \ctx -> do
-        sourceWallet <- emptyRandomWallet ctx
-
-        r <- request @[ApiTransaction n] ctx
-            (Link.migrateWallet @'Byron sourceWallet)
-            Default
-            (NonJson "{passphrase:,}")
-        expectResponseCode @IO HTTP.status400 r
-        expectErrorMessage errMsg400ParseError r
-
     it "BYRON_GET_02 - Byron ep does not show Shelley wallet" $ \ctx -> do
         w <- emptyWallet ctx
         r <- request @ApiByronWallet ctx
@@ -1321,268 +1267,6 @@ spec = do
         expectResponseCode @IO HTTP.status404 r
         expectErrorMessage (errMsg404NoWallet $ w ^. walletId) r
 
-    it "BYRON_CALCULATE_03 - \
-        \Cannot estimate migration for Shelley wallet using Byron endpoint"
-        $ \ctx -> do
-            w <- emptyWallet ctx
-            let ep = Link.getMigrationInfo @'Byron w
-            r <- request @ApiWalletMigrationInfo ctx ep Default Empty
-            expectResponseCode @IO HTTP.status404 r
-            expectErrorMessage (errMsg404NoWallet $ w ^. walletId) r
-
-    it "BYRON_MIGRATE_01 - \
-        \after a migration operation successfully completes, the correct \
-        \amount eventually becomes available in the target wallet for arbitrary \
-        \ number of specified addresses."
-        $ \ctx -> do
-              testAddressCycling ctx 1
-              testAddressCycling ctx 3
-              testAddressCycling ctx 10
-
-    it "BYRON_MIGRATE_01 - \
-        \ migrate a big wallet requiring more than one tx" $ \ctx -> do
-        pendingWith "Byron wallets not support in cardano-node yet"
-        -- NOTE
-        -- Special mnemonic for which 500 legacy funds are attached to in the
-        -- genesis file.
-        --
-        -- Out of these 500 coins, 100 of them are of 1 Lovelace and are
-        -- expected to be treated as dust. The rest are all worth:
-        -- 10,000,000,000 lovelace.
-        let mnemonics =
-                ["collect", "fold", "file", "clown"
-                , "injury", "sun", "brass", "diet"
-                , "exist", "spike", "behave", "clip"
-                ] :: [Text]
-        let payloadRestore = Json [json| {
-                "name": "Big Byron Wallet",
-                "mnemonic_sentence": #{mnemonics},
-                "passphrase": #{fixturePassphrase},
-                "style": "random"
-                } |]
-        (_, wOld) <- unsafeRequest @ApiByronWallet ctx
-            (Link.postWallet @'Byron) payloadRestore
-        eventually "wallet balance greater than 0" $ do
-            request @ApiByronWallet ctx
-                (Link.getWallet @'Byron wOld)
-                Default
-                Empty >>= flip verify
-                [ expectField (#balance . #available) (.> Quantity 0)
-                ]
-        let originalBalance = view (#balance . #available . #getQuantity) wOld
-
-        -- Calculate the expected migration fee:
-        rFee <- request @ApiWalletMigrationInfo ctx
-            (Link.getMigrationInfo @'Byron wOld)
-            Default
-            Empty
-        verify rFee
-            [ expectResponseCode @IO HTTP.status200
-            , expectField #migrationCost (.> Quantity 0)
-            ]
-        let expectedFee = getFromResponse (#migrationCost . #getQuantity) rFee
-
-        -- Migrate to a new empty wallet
-        wNew <- emptyWallet ctx
-        addrs <- listAddresses @n ctx wNew
-        let addr1 = (addrs !! 1) ^. #id
-
-        let payloadMigrate =
-                Json [json|
-                    { passphrase: #{fixturePassphrase}
-                    , addresses: [#{addr1}]
-                    }|]
-        request @[ApiTransaction n] ctx
-            (Link.migrateWallet @'Byron wOld)
-            Default
-            payloadMigrate >>= flip verify
-            [ expectResponseCode @IO HTTP.status202
-            , expectField id ((`shouldBe` 2). length)
-            ]
-
-        -- Check that funds become available in the target wallet:
-        let expectedBalance = originalBalance - expectedFee
-        eventually "wallet balance = expectedBalance" $ do
-            request @ApiWallet ctx
-                (Link.getWallet @'Shelley wNew)
-                Default
-                Empty >>= flip verify
-                [ expectField
-                        (#balance . #getApiT . #available)
-                        ( `shouldBe` Quantity expectedBalance)
-                , expectField
-                        (#balance . #getApiT . #total)
-                        ( `shouldBe` Quantity expectedBalance)
-                ]
-
-        -- Analyze the target wallet UTxO distribution
-        request @ApiUtxoStatistics ctx (Link.getUTxOsStatistics @'Shelley wNew)
-            Default
-            Empty >>= flip verify
-            [ expectField
-                #distribution
-                ((`shouldBe` (Just 400)) . Map.lookup 10_000_000_000)
-            ]
-
-    it "BYRON_MIGRATE_01 - \
-        \a migration operation removes all funds from the source wallet."
-        $ \ctx -> forM_ [fixtureRandomWallet, fixtureIcarusWallet]
-        $ \fixtureByronWallet -> do
-            pendingWith "Byron wallets not supported in cardano-node yet."
-            -- Restore a Byron wallet with funds, to act as a source wallet:
-            sourceWallet <- fixtureByronWallet ctx
-
-            -- Perform a migration from the source wallet to a target wallet:
-            targetWallet <- emptyWallet ctx
-            addrs <- listAddresses @n ctx targetWallet
-            let addr1 = (addrs !! 1) ^. #id
-
-            r0 <- request @[ApiTransaction n] ctx
-                (Link.migrateWallet @'Byron sourceWallet)
-                Default
-                (Json [json|
-                    { passphrase: #{fixturePassphrase}
-                    , addresses: [#{addr1}]
-                    }|])
-            verify r0
-                [ expectResponseCode @IO HTTP.status202
-                , expectField id (`shouldSatisfy` (not . null))
-                ]
-
-            -- Verify that the source wallet has no funds available:
-            r1 <- request @ApiByronWallet ctx
-                (Link.getWallet @'Byron sourceWallet) Default Empty
-            verify r1
-                [ expectResponseCode @IO HTTP.status200
-                , expectField (#balance . #available) (`shouldBe` Quantity 0)
-                ]
-
-    it "BYRON_MIGRATE_02 - \
-        \migrating an empty wallet should fail."
-        $ \ctx -> forM_ [emptyRandomWallet, emptyIcarusWallet] $ \emptyByronWallet -> do
-            sourceWallet <- emptyByronWallet ctx
-            targetWallet <- emptyWallet ctx
-            addrs <- listAddresses @n ctx targetWallet
-            let addr1 = (addrs !! 1) ^. #id
-            let payload =
-                    Json [json|
-                        { passphrase: #{fixturePassphrase}
-                        , addresses: [#{addr1}]
-                        }|]
-            let ep = Link.migrateWallet @'Byron sourceWallet
-            r <- request @[ApiTransaction n] ctx ep Default payload
-            let srcId = sourceWallet ^. walletId
-            verify r
-                [ expectResponseCode @IO HTTP.status403
-                , expectErrorMessage (errMsg403NothingToMigrate srcId)
-                ]
-
-    it "BYRON_MIGRATE_02 - \
-        \migrating wallet with dust should fail."
-        $ \ctx -> do
-            -- NOTE
-            -- Special mnemonic for which wallet with dust
-            -- (5 utxos with 60 lovelace in total)
-            let mnemonics =
-                    [ "suffer", "decorate", "head", "opera"
-                    , "yellow", "debate", "visa", "fire"
-                    , "salute", "hybrid", "stone", "smart"
-                    ] :: [Text]
-            let payloadRestore = Json [json| {
-                    "name": "Dust Byron Wallet",
-                    "mnemonic_sentence": #{mnemonics},
-                    "passphrase": #{fixturePassphrase},
-                    "style": "random"
-                    } |]
-            pendingWith "Byron wallets not supported in cardano-node yet."
-            (_, sourceWallet) <- unsafeRequest @ApiByronWallet ctx
-                (Link.postWallet @'Byron) payloadRestore
-            eventually "wallet balance greater than 0" $ do
-                request @ApiByronWallet ctx
-                    (Link.getWallet @'Byron sourceWallet)
-                    Default
-                    Empty >>= flip verify
-                    [ expectField (#balance . #available) (.> Quantity 0)
-                    ]
-
-            targetWallet <- emptyWallet ctx
-            addrs <- listAddresses @n ctx targetWallet
-            let addr1 = (addrs !! 1) ^. #id
-            let payload =
-                    Json [json|
-                        { passphrase: #{fixturePassphrase}
-                        , addresses: [#{addr1}]
-                        }|]
-            let ep = Link.migrateWallet @'Byron sourceWallet
-            r <- request @[ApiTransaction n] ctx ep Default payload
-            let srcId = sourceWallet ^. walletId
-            verify r
-                [ expectResponseCode @IO HTTP.status403
-                , expectErrorMessage (errMsg403NothingToMigrate srcId)
-                ]
-
-    it "BYRON_MIGRATE_03 - \
-        \actual fee for migration is the same as the predicted fee."
-        $ \ctx -> forM_ [fixtureRandomWallet, fixtureIcarusWallet]
-        $ \fixtureByronWallet -> do
-            pendingWith "Byron wallets not supported in cardano-node yet."
-            -- Restore a Byron wallet with funds.
-            sourceWallet <- fixtureByronWallet ctx
-
-            -- Request a migration fee prediction.
-            let ep0 = (Link.getMigrationInfo @'Byron sourceWallet)
-            r0 <- request @ApiWalletMigrationInfo ctx ep0 Default Empty
-            verify r0
-                [ expectResponseCode @IO HTTP.status200
-                , expectField #migrationCost (.> Quantity 0)
-                ]
-
-            -- Perform the migration.
-            targetWallet <- emptyWallet ctx
-            addrs <- listAddresses @n ctx targetWallet
-            let addr1 = (addrs !! 1) ^. #id
-            let payload =
-                    Json [json|
-                        { passphrase: #{fixturePassphrase}
-                        , addresses: [#{addr1}]
-                        }|]
-            let ep1 = Link.migrateWallet @'Byron sourceWallet
-            r1 <- request @[ApiTransaction n] ctx ep1 Default payload
-            verify r1
-                [ expectResponseCode @IO HTTP.status202
-                , expectField id (`shouldSatisfy` (not . null))
-                ]
-
-            -- Verify that the fee prediction was correct.
-            let actualFee = fromIntegral $ sum $ apiTransactionFee
-                    <$> getFromResponse id r1
-            let predictedFee =
-                    getFromResponse (#migrationCost . #getQuantity) r0
-            actualFee `shouldBe` predictedFee
-
-    it "BYRON_MIGRATE_04 - migration fails with a wrong passphrase"
-        $ \ctx -> forM_ [fixtureRandomWallet, fixtureIcarusWallet]
-        $ \fixtureByronWallet -> do
-        pendingWith "Byron wallets not supported in cardano-node yet."
-        -- Restore a Byron wallet with funds, to act as a source wallet:
-        sourceWallet <- fixtureByronWallet ctx
-
-        -- Perform a migration from the source wallet to a target wallet:
-        targetWallet <- emptyWallet ctx
-        addrs <- listAddresses @n ctx targetWallet
-        let addr1 = (addrs !! 1) ^. #id
-        r0 <- request @[ApiTransaction n] ctx
-            (Link.migrateWallet @'Byron sourceWallet)
-            Default
-            (Json [json|
-                { passphrase: "not-the-right-passphrase"
-                , addresses: [#{addr1}]
-                }|])
-        verify r0
-            [ expectResponseCode @IO HTTP.status403
-            , expectErrorMessage errMsg403WrongPass
-            ]
-
     it "NETWORK_SHELLEY - Wallet has the same tip as network/information" $
         \ctx -> do
             let getNetworkInfo = request @ApiNetworkInformation ctx
@@ -1608,21 +1292,6 @@ spec = do
                     , expectField (#tip . #height) (`shouldBe` blockHeight)
                     ]
   where
-    -- Compute the fee associated with an API transaction.
-    apiTransactionFee :: ApiTransaction n -> Word64
-    apiTransactionFee t =
-        inputBalance t - outputBalance t
-      where
-        inputBalance = fromIntegral
-            . sum
-            . fmap (view (#amount . #getQuantity))
-            . mapMaybe ApiTypes.source
-            . view #inputs
-        outputBalance = fromIntegral
-            . sum
-            . fmap (view (#amount . #getQuantity))
-            . view #outputs
-
     genMnemonics
        :: forall mw ent csz.
            ( ConsistentEntropy ent mw csz
@@ -1634,55 +1303,3 @@ spec = do
        => IO [Text]
     genMnemonics =
         mnemonicToText . entropyToMnemonic @mw <$> genEntropy
-
-    testAddressCycling ctx addrNum =
-        forM_ [fixtureRandomWallet, fixtureIcarusWallet]
-        $ \fixtureByronWallet -> do
-            pendingWith "Byron wallets not supported in cardano-node yet."
-            -- Restore a Byron wallet with funds, to act as a source wallet:
-            sourceWallet <- fixtureByronWallet ctx
-            let originalBalance =
-                        view (#balance . #available . #getQuantity) sourceWallet
-
-            -- Create an empty target wallet:
-            targetWallet <- emptyWallet ctx
-            addrs <- listAddresses @n ctx targetWallet
-            let addrIds =
-                    map (\(ApiTypes.ApiAddress theid _) -> theid) $
-                    take addrNum addrs
-
-            -- Calculate the expected migration fee:
-            r0 <- request @ApiWalletMigrationInfo ctx
-                (Link.getMigrationInfo @'Byron sourceWallet) Default Empty
-            verify r0
-                [ expectResponseCode @IO HTTP.status200
-                , expectField #migrationCost (.> Quantity 0)
-                ]
-            let expectedFee = getFromResponse (#migrationCost . #getQuantity) r0
-
-            -- Perform a migration from the source wallet to the target wallet:
-            r1 <- request @[ApiTransaction n] ctx
-                (Link.migrateWallet @'Byron sourceWallet)
-                Default
-                (Json [json|
-                    { passphrase: #{fixturePassphrase}
-                    , addresses: #{addrIds}
-                    }|])
-            verify r1
-                [ expectResponseCode @IO HTTP.status202
-                , expectField id (`shouldSatisfy` (not . null))
-                ]
-
-            -- Check that funds become available in the target wallet:
-            let expectedBalance = originalBalance - expectedFee
-            eventually "Wallet has expectedBalance" $ do
-                r2 <- request @ApiWallet ctx
-                    (Link.getWallet @'Shelley targetWallet) Default Empty
-                verify r2
-                    [ expectField
-                            (#balance . #getApiT . #available)
-                            (`shouldBe` Quantity expectedBalance)
-                    , expectField
-                            (#balance . #getApiT . #total)
-                            (`shouldBe` Quantity expectedBalance)
-                    ]
